+++ picklepipe/__init__.py
@@ -8,16 +8,18 @@ from .pipe import (BaseSerializingPipe,
                    make_pipe_pair)
 from .picklepipe import PicklePipe
 from .marshalpipe import MarshalPipe
+from .jsonpipe import JSONPipe
 
 __author__ = 'Seth Michael Larson'
 __email__ = 'sethmichaellarson@protonmail.com'
 __license__ = 'MIT'
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 
 __all__ = [
     'BaseSerializingPipe',
     'PicklePipe',
     'MarshalPipe',
+    'JSONPipe',
     'PipeClosed',
     'PipeError',
     'PipeTimeout',